Prevent VMI_ERROR for FILE mode.
The methods for finding config values that require
virtual translations will not work for FILE mode,
so skips them to prevent the VMI_ERROR from being
printed.

Also prevents the suggestion for setting win_kdvb
from being printed in some cases where it is
present in the config, but was not enough to be
useful because of FILE mode or some other value
was missing.

Attempts to clear up how various combinations
of config values can be used and sanity checks
them against each other as much as possible.