Add checkpatch.pl from Linux/Qemu into repo
Moving forward, I'd like to have some automated checking on the style
for new patches.  The idea is to keep things more consistent and
readable in the code base.  Much of the open source C development
community is familiar with using checkpatch.pl, so I propose that we
adopt this style as well.

After including this in the repository, I can put together some Jenkins
jobs to test the style of PRs.