Fix KASLR offset bug for x86_64 (#600)
When the kernel version is 4.13.0+, the previous kaslr lookup function
fails somehow.

Determining KASLR offset in x86_64 can be done in a simpler way following
the official kernel address layout documentation:
    https://www.kernel.org/doc/Documentation/x86/x86_64/mm.txt

What we need to do is just keep attemping to get the k2p mapping of the first
page mapped in kernel text mapping. We can get the offset once we get the first
kernel text page mapped. Because in the range of 0xffffffff80000000 ~
0xffffffffa0000000 there is only kernel text, and the first page found mapped
beginning from 0xffffffff81000000 is the start of the kernel text mapping, thus
we can get the kaslr offset.

Note that here it should be 0xffffffff81000000 instead of 0xffffffff80000000.

Also, kernel text is by default using 2M huge page so we are adding 0x200000
instead of 0x1000.

Tested on x86_64 Ubuntu 16.04 with kernel v4.4.0 and v4.15.0.