Add domain watcher
This patch adds a domain watcher that signals a new domain or if a domains was destroyed. The callback function returns a structure with domain id, uuid and if the domain was created or not.

After a domain is created it is inserted in a gTree so we can have
information about the destroy event.

Signed-off-by: Alexandru Isaila <aisaila@bitdefender.com>