bug fix: set HAVE_LIBVMI_REQUEST ON while find it
The variable in subdirectories cannot be found in parent subdirectories. That is, even if LibvmiRequest is found, HAVE_LIBVMI_REQUEST cannot be defined in config.h.
It is a first-aid fix and the coding in cmake is not very pretty after the fix. I think the /CMakeLists.txt , /libvmi/CMakeLists.txt , and /cmake/modules/*.cmake should be optimized throughly in the future.