Fatal bug fix about cmake (#941)
* bug fix: set HAVE_LIBVMI_REQUEST ON while find it

The variable in subdirectories cannot be found in parent subdirectories. That is, even if LibvmiRequest is found, HAVE_LIBVMI_REQUEST cannot be defined in config.h.
It is a first-aid fix and the coding in cmake is not very pretty after the fix. I think the /CMakeLists.txt , /libvmi/CMakeLists.txt , and /cmake/modules/*.cmake should be optimized throughly in the future.

* fix it by passing

* bug fix: add '-lm'

In kvm_legacy.c, there is a function 'ceil', which needs the library 'libm'. 
It seems that, without HAVE_LIBVMI_REQUEST, the branch to the function 'ceil' will be optimized by gcc. When I use cmake to test kvm_legacy with HAVE_LIBVMI_REQUEST(after I fix another bug), I meet the error with unreferenced symbol 'ceil'. To add the link library 'libm' can resolve it.