enable the kvm_legacy to fetch multicpu reg info (#994)
* bug fix: set HAVE_LIBVMI_REQUEST ON while find it

The variable in subdirectories cannot be found in parent subdirectories. That is, even if LibvmiRequest is found, HAVE_LIBVMI_REQUEST cannot be defined in config.h.
It is a first-aid fix and the coding in cmake is not very pretty after the fix. I think the /CMakeLists.txt , /libvmi/CMakeLists.txt , and /cmake/modules/*.cmake should be optimized throughly in the future.

* fix it by passing

* bug fix: add '-lm'

In kvm_legacy.c, there is a function 'ceil', which needs the library 'libm'. 
It seems that, without HAVE_LIBVMI_REQUEST, the branch to the function 'ceil' will be optimized by gcc. When I use cmake to test kvm_legacy with HAVE_LIBVMI_REQUEST(after I fix another bug), I meet the error with unreferenced symbol 'ceil'. To add the link library 'libm' can resolve it.

* Create kvm-qemu-v2.12-libvmi.patch

* enable the kvm_legacy to fetch multicpu reg info

In kvm_legacy mode, the vmi_get_vcpureg API can only fetch reg info with cpu 0 because the inner implement unuse the argument "vcpu".
The patch make up the flaw.

* undo the mistake change 

The commit is accidently done. Undo it.